requirements.txt
# The order of packages is significant, because pip processes them in the order
# of appearance. Changing the order has an impact on the overall integration
# process, which may cause wedges in the gate later.
pbr!=0.7,<1.0,>=0.6

Paste
PasteDeploy>=1.5.0
Routes!=2.0,>=1.12.3
eventlet!=0.17.0,>=0.16.1
greenlet>=0.3.2
httplib2>=0.7.5
requests!=2.4.0,>=2.2.0
jsonrpclib
Jinja2>=2.6 # BSD License (3 clause)
keystonemiddleware<1.6.0,>=1.5.0
netaddr>=0.7.12
python-neutronclient<2.5.0,>=2.4.0
retrying!=1.3.0,>=1.2.3 # Apache-2.0
SQLAlchemy<=0.9.99,>=0.9.7
WebOb>=1.2.3
python-keystoneclient<1.4.0,>=1.2.0
alembic<0.8.1,>=0.7.2
six>=1.9.0
stevedore<1.4.0,>=1.3.0 # Apache-2.0
oslo.concurrency<1.9.0,>=1.8.2 # Apache-2.0
oslo.config<1.10.0,>=1.9.3 # Apache-2.0
oslo.context<0.3.0,>=0.2.0 # Apache-2.0
oslo.db<1.8.0,>=1.7.0 # Apache-2.0
oslo.i18n<1.6.0,>=1.5.0 # Apache-2.0
oslo.log<1.1.0,>=1.0.0 # Apache-2.0
oslo.messaging<1.9.0,>=1.8.0 # Apache-2.0
oslo.middleware<1.1.0,>=1.0.0 # Apache-2.0
oslo.rootwrap<1.7.0,>=1.6.0 # Apache-2.0
oslo.serialization<1.5.0,>=1.4.0 # Apache-2.0
oslo.utils<1.5.0,>=1.4.0 # Apache-2.0

python-novaclient<2.24.0,>=2.22.0

taskflow>=0.1.3,<=0.3.21
